virt: move minimum balloon size in Balloon class
Make the minimum balloon size a real Balloon device property,
and use that as authoritative source to report it.
Previously, we just computed from the vm.conf.

+++

Our modelling of the Balloon device is debatable, and sometimes
awkward. In particular, it is debatable that the balloon device
should keep its state, while we modify it using Domain methods
(look at libvirt methods invoked under the hood).

However, what's surely the bigger evil is to have the state spread
across the Vm class and the Balloon device.
Until we find a better modeling, we have now a complete and consistent
Balloon device. As minor bonus, we go one babystep forward in the task
of slimming down the Vm class.

Change-Id: I487221c001c292a0445909d9255cbbc1851c433a
Signed-off-by: Francesco Romani <fromani@redhat.com>