virt: add method to build device map from XML
We already have one helper module to create a device map
out of vm.conf; this patch add a companion method
to build a device map from a domain XML.

Please note that this method is just about building objects out
of the domain XML we receive, we don't want to handle the filtering
of unknown devices here.

We also simplify a bit the identification of the devices.
There is no real need for a blacklist of devices; if it is outside
our mapping, we can ignore the data and not build one object out of it.

Change-Id: I86c8da1bbc2912f1e2112f348f8b2dafa811e8c7
Signed-off-by: Francesco Romani <fromani@redhat.com>