virt: Switch to persistent domains
We have been using transient domains for running VM so far.  They disappear
when their VMs stop running at the given host.  Detailed information about a
destroyed VM is still retained in its recovery file, until Engine asks for
releasing it.

We would like to stop using Vdsm specific recovery files and use libvirt domain
XML as the complete and ultimate source of information about a VM instead.  In
order to be able to access the domain data before the VM is started or after
it's stopped, similarly to the recovery files, we need persistent domains that
don't disappear immediately.

This patch switches from transient to persistent libvirt domains.  Creation and
removal of the domains is handled similarly to recovery files.  The only extra
feature is that we must remove the domain, if it exists, before a VM is
defined.

This patch has already been applied and reverted, due to missing support in
libvirt for live storage migration of persistent domains (see
https://bugzilla.redhat.com/1458548 and https://bugzilla.redhat.com/1459113).
The support is available in development libvirt versions now, so we can apply
the patch again with a required change to blockCopy libvirt call and test it.
We keep backward compatibility with released libvirt versions until the updated
libvirt is generally available.

Change-Id: Ia9ddc00515708b5c04f0b585aef4c39b0a67e58c
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>