create-snapshot: Use unsafe option when creating snapshots
With new qemu-img support for unsafe operations (added in qemu-img
2.10.0), and as the VM may be running while creating snapshots,
we have to use new qemu-img support for unsafe operations. When
using the new unsafe support, qemu-img doesn't access the backing file,
hence we have to provide the size.

Providing the size of the volume will save preparing and tearing down
the volume, thus improving the performance of snapshot creation.

However, when working with qemu-img < 2.10, we do have to prepare and
the backing volume despite the fact that we add the size, because it is
NOT documented that qemu-img will not access the backing chain.

Change-Id: If7525e5f3d9e18eaf20aefd4ce25e8dcc48ee675
Bug-Url: https://bugzilla.redhat.com/1395941
Signed-off-by: Ala Hino <ahino@redhat.com>