virt: tests: unify listAllDevices{,2}
We currently have two faked listAllDevices, one of which is used
only for the hostdev performance test.

A test run under python3 revelead issues with our monkeypatch
implementation, most likely related to confusion between instance/classmethod
and subtle changes to object model between python2 and python3, like
we faced in Ie3100455ab761481f57a93ab8aacc34ac4ab91eb

The actual source of error is yet to be understood, but we can just
sidestep this issue by removing some duplication, and we do that
in this patch.

Change-Id: Id682fdcc319a8050de54024c569625090dea6630
Signed-off-by: Francesco Romani <fromani@redhat.com>