virt: metadata: introduce the Descriptor class
The device-specific metadata is stored in the domain metadata, as
children of the main domain metadata element. The same applies for
the custom properties. Those two cases are the only two on which
the nesting is allowed on metadata.

A full-featured and legal domain metadata layout could look
like (namespace omitted for brevity)

  <metadata>
    <vm>
      <foo>1</foo>
      <bar>2</bar>
      <custom>
        <fizz>a</fizz>
      </custom>
      <device alias='dev0'>
        <buzz>b</buzz>
      </device>
    </vm>
  </metadata>

the metadata.domain() context manager allows to modify only the VM
metadata, so it must returns top-level properties.
It should thus ignore the per-device custom properties, and the VM
custom properties. Currently, if we write back the above metadata
we will get

  <metadata>
    <vm>
      <foo>1</foo>
      <bar>2</bar>
      <custom />
      <device />
    </vm>
  </metadata>

This happens because the current code goes too far ignoring the nested data,
and wipes them out when the metadata is rewritten.
This is because actually both domain() and device() context manager need
to keep too much implicit state and fail to do so.

So, the root cause is the implicit state and its management. To solve
this issue, we choose to remove any implicit state management and
we introduce one explicit Descriptor class, which will keep the current
metadata and allow the client code to do fine-grained management.

the Descriptor class will replace:
1. metadata.from_xml()
   use metadata.Descriptor.from_xml() instead
2. metadata.domain()
   use metadata.Descriptor.load()/values()/dump() instead
3. metadata.device()
   use the `device` context manager of metadata.Descriptor instead
4. metadata.device_from_xml_tree()
   use metadata.Descriptor.from_xml() and the `device` context manager
   instead

Future patches will port the new code to the new Descriptor class,
and later we will remove the obsolete functions.

Change-Id: I0d60c32ba4663538d870bce9b7cca7adcb117a07
Signed-off-by: Francesco Romani <fromani@redhat.com>