blockSD: Add create_volume_lease() method
Move the code for creating a volume lease to blockSD, since this is part
of the block storage domain implementation. Moving this code removes a
circular dependency between blockVolume and sd modules.

The RESERVED_LEASES constant was moved to blockSD since it is an
implementation detail of block storage domain, and needed for creating a
new lease.

Moving the code revealed duplicate code for calculating a volume lease
offset. The duplicate code was extracted to new volume_lease_offset()
method.

Change-Id: I60a9966dd5e0b35933bc01f6e76d2b0573985d70
Signed-off-by: Nir Soffer <nsoffer@redhat.com>