stomp: implement client reconnect
After enabling incoming heartbeats in AsyncDispatcher,
an implementation of handle_timeout is presented in the
client side also.
In the client side we wish to reconnect when we stop receiving
incoming heartbeats according to a retry policy which defines
number of tries and the timeout we wish to wait.
In order to make sure incoming heartbeats are received and we won't
close the socket before getting them, we add grace period for incoming
heartbeats and reduce the outgoing heartbeats interval in order to make
sure the outgoing heartbeats are received in server side.
By default, number of retries after timeout is 1 to ensure fast recovery.

Change-Id: I838edc76f8904a98a6a220b44a4e5752d9e61050
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1376843