virt: Make sure VMs are restored as paused
When we restore VMs from snapshot, we expect that they are restored as paused
and we then unpause them with `resume' call in Vm._underlyingCont.  However at
least with newer libvirt versions the VMs may be actually run automatically
from `restore' call.  Then the `resume' call fails and the VM is lost.

It's probably change from former libvirt behavior, however the libvirt
documentation says: "Normally, the saved state file will remember whether the
domain was running or paused, and restore defaults to the same state."  So the
current behavior can't be classified as unexpected or even buggy.

libvirt offers flags to specify whether to restore the VM as paused or running.
Let's use VIR_DOMAIN_SAVE_PAUSED flag as the least invasive fix guaranteeing
consistent behavior.

Bug-Url: https://bugzilla.redhat.com/1461811
Backport-To: 4.1
Change-Id: I4397516d56ce401a886294c0e1d8c892687f9144
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>