tests: Add storage_qcow2_test to test modules
This test was never added to the makefile, so it never run in the CI.

Running on travis show that some tests are broken since the file system
on travis does not support sparseness. This issue was discovered before
when we tested qemuimg.map(). Mark the broken tests as broken on travis.

Change-Id: Ic25556732ac9a52b7b5fdc2af5a0a297511be95f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>