virt: devices: simplify 'device' and 'type'
The vmdevices.core.parse_device_params() tried to be a generic
helper to parse the very basic attributes common to all the devices.

Among other things, it tried to extract the 'device' and 'type'
attributes, used all across our device class hierarchy.

We should settle for one meaning of those attributes:
- device: the device class (e.g. Sound, Tpm, Graphics...)
- type: the device subtype, should the device have one.

The problem is that this is never ever enforced, so each and every class
has the complete freedom to use those attributes how it likes.
So, the 'type' could not be used at all, or used as alias for 'device',
or any theoretical combination. And we had quite a lot of those
combinations actually used in the code.

The parse_device_params() helper tried to generalize a bit and
return the aforementioned mapping, but this backfired badly,
and we had a proliferation of reassignment or swaps between
params['device'] and params['type']. This was utterly confusing.

A cleanup of this mess is out of the scope for the devices-from-xml
tree. The best we can do now is remove the failed attempt to generalize,
and do the mapping per-device basis.
At least now the code is less obscure, and we don't further increase the
mess we already have.

Change-Id: I4fd3fcaa96db11d5b3f1345d9afcb8117090fba0
Signed-off-by: Francesco Romani <fromani@redhat.com>