vm: create devices from engine XML
Now that most of the devices allow to be initialized
from their XML representation, we can support the engine XML creation
flow. This patch does so adding the final pieces to the puzzle.

In the existing, vm.conf-based creation flow, Vdsm will
adjust the device configuration received by Engine on per-host basis
(example: leases, storage, OVS networking...), then initialize
the devices object from that configuration. The last step is to
compose the libvirt domain XML using the device objects and the
VM configuration received by Engine.

The XML creation flow is slightly different. Engine sends almost all
the VM configuration already encoded in a well-formed domain XML,
with the notable exceptions of the storage devices (see below).
Vdsm parses the information it needs from this XML to fill the VM and
devices data structures, but most of the XML is unchanged. The
only changes are again the host-specific parts, which Vdsm fills
with the data it gathered.

Actually, currently Vdsm doesn't fill the gaps, but rather regenerates
the device-specific XML snippet. The end result is the same, but we
took this approach because it is the one which requires less changes
considering the current infrastructure.

The only notable exception is the storage devices. Here Vdsm still
expects the configuration to be sent *only* in the legacy vm.conf format,
and will rebuild the devices XML snippets from scratch as it did before.
We choose this approach because the high complexity of the storage
devices.

Change-Id: I315d850c089dd48b01fbd2d7af174fe1ce8ab10d
Signed-off-by: Francesco Romani <fromani@redhat.com>