tests: Fix mailbox thread leak test
The thread leak test was added to make sure we close the thread pool and
wait for the threads when stopping the mail monitor.

There are many issues in this test:
- Not sure what the asserts were testing. Looking at the original bug,
  we should test that the worker pool threads were stopped, but the test
  was asserting that the number of threads is the same as it was when
  running the mailer.
- The test was invoking run() - this can never happen in the real
  code and is invalid use of this code. I think this was an attempt to
  deal with stop being asynchronous.
- The test was using utils.retry - this helper does not work in pytest

Fix all the issues by simplifying the test to count the number of threads
before creating the mail monitor, stop and wait until the mail monitor
was stopped, and check that we did not leak any threads.

Change-Id: I67aaf03e382c730757e8d3d208267d177ea325c6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>