tests: Make permutations pytest friendly
pytest does not support the character "[" in a permuted test name. It
assumes that the means the test was parameterized, and remove the
character when collecting methods, breaking collection.

Fix the issue by replacing "[" and "]" to "(" and ")" in test names.
With this change, all storage tests pass with pytest on python 2.7, and
about 30% of the tests pass on python 3.5.

Change-Id: I7f6bdf5a3c7c4ce2dc214e4b5cb589dd36b3a592
Signed-off-by: Nir Soffer <nsoffer@redhat.com>