python3: Make missing sanlock module pass import
On python 3 we are using a special missing sanlock module, that will
fail when used, so we can test code monkeypatching sanlock on python 3.
However sanlock constants are used during import, failing import of all
the tests that import sanlock.

Add the missing constants to sanlock, so importing the tests works.
These tests can now be marked with pytestmark so they are skipped on
python 3.

Change-Id: I8dd19c3b2156a3195ea945fb18bbfa5003781148
Signed-off-by: Nir Soffer <nsoffer@redhat.com>