python3: Fix UUID packing/unpacking on python 3
The old code was broken on python 3 because it used float division (/),
and was depending on long integer representation ("L" suffix).  It also
tried to reinvent the uuid module. We use now the uuid module for
converting to/from integer value and uuid string.

Due to PyCQA/pylint#961, we have to disable pylint member check when
accessing UUID's int attribute.

Change-Id: I41cfa5e54e76dfcad267792daebc54e80bf2a022
Signed-off-by: Nir Soffer <nsoffer@redhat.com>