virt: Don't call destroy on destination after failed migration
When a migration fails, the source host calls destroy on the destination host.
That's not necessary, since libvirt destroys the VM itself (and the call
swallows any exception).  Handling several destroy calls on the destination
used to be buggy and has been fixed.  Destroy on the migration destination
should be invoked only by libvirt or Engine, that's enough.

This patch removes the superfluous destroy call from the migration source.

Change-Id: I0d0bff00352c28e937a8514aa7b6ec49b7121c85
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>