virt: hostdev: add device_name_from_address
For USB devices, it is not safe to assume that
device name is always in the form "usb_%{bus}_%{device}".
There are multiple cases in the wild on which this naive
assumption doesn't hold.

However, we already retrieve the full device listing
from libvirt, and among other information we have the device name and
its address, so we can build the mapping we need to
later safely retrieve the device name given its address.

Change-Id: I0cef83c23a13b0d15d6bace26572bf5a8908ce7c
Signed-off-by: Francesco Romani <fromani@redhat.com>