lib: add collectd client
add collectd client library, using the plain text protocol
over the unixsock plugin.

We ship the configuration for the unixsock plugin; however,
it must be noted that this protocol is not well suited to
transfer large amount of data. We may need to switch to a more
performant solution to handle high loads.

However, the end goal is to make Engine consume data directly
from collectd (or the final metrics store), so this is
a good first step and backward compatibility option.

NOTE: until rhbz#1304029 is fixed, we will still need manual
intervention.

NOTE: Unfortunately, this code:
https://github.com/collectd/collectd/blob/master/contrib/collectd_unixsock.py
is not shipped by default (nor, as far as we know, supported
as official python client), so we roll our own.

Change-Id: I2bf83e3bcc42aa114b178274f9172d8bd53dfd1c
Signed-off-by: Francesco Romani <fromani@redhat.com>