stomp: set default heartbeat values and add grace period
default incoming heartbeat used to be 5 seconds but was never checked.
In an actual system the heartbeat should be greater, since CPU and network
spikes affect the server communication and don't indicate there is a
real communication problem.
Also, in order to make sure incoming heartbeats are received and we won't
close the socket before getting them, we add grace period for incoming
heartbeats and reduce the outgoing heartbeats interval in order to make
sure the outgoing heartbeats are received in server side.
By default, number of retries after timeout is 1 to ensure fast recovery.

Change-Id: I0e8c4f2fef9083521dcf2da12cf1f9db84b75492
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1376843