vm: Log the time took to extend a volume
Measure the time to extend a volume from the time we decided that the
volume should be extended, until the time the volume is refreshed after
the extension.

Note that during live storage migration we extend the replica volume
before extending the original volume, so the time to extend the original
volume includes the time to extend the replica.

Change-Id: Ie970247f7694da644e4789b674884ceb8e4210be
Signed-off-by: Nir Soffer <nsoffer@redhat.com>