virt: recovery: update domain XML from libvirt
In the XML creation flow, we take steps to make sure we save the
domain metadata in the libvirt domain XML. We do not, however,
alter in any way the XML we receive from the client (Engine).
This is a good practice, because we should consider the VM params
as read-only.

However, this means that the client-provided domain XML is no longer
up to date in the recovery; the one stored in libvirt always is,
so in the recovery path we should override the XML stored in the
recovery file with a fresh copy obtained from libvirt.

Change-Id: I0e4f278487fd4e7063d7c14fcf0cfa108a20eeaf
Signed-off-by: Francesco Romani <fromani@redhat.com>