recovery: Initial support for external VMs
It's useful to show users external VMs running on their hosts.  Engine already
recognizes external VMs and handles them appropriately.  So all we need to do
is to report external VMs to Engine.

This patch reads in external VMs on recovery.  We should watch for external VMs
periodically, which functionality will be provided in a followup patch.

Change-Id: I8307860965be1ca21c3fab0364f78c15aed7aa91
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>