v2v: allow import of disks not in storage pool
It should be possible to import also disks that are not in libvirt
storage pool. Previously we required that all local disks
(<disk type='file' ...>) exist in some storage pool. This requirement is
unjustified and having no storage pools at all is completely valid use
case.

The only exception is Xen source where virDomainGetBlockInfo() is not
implemented.

Change-Id: Id0562848e1eb655e8426563ea5d161ea790aaafb
Bug-Url: https://bugzilla.redhat.com/1468944
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>