v2v: allow import of disk with type='volume'
Currently we support only import of disks with type either 'file' or
'block'. In libvirt, it is possible to have a disk with type='volume' in
which case the name of the storage pool and volume name are then
supplied in domain XML.

This is mostly useful for KVM source. Theoreticaly it could be also used
with Xen sources. But we support Xen only on RHEL5 and 'volume' type is
not available there (libvirt is too old).

Change-Id: I4d6afdfa8465c40eca110e87738e30a975f176a1
Bug-Url: https://bugzilla.redhat.com/1469077
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>