check: Handle errors the complete callback
When a check is completed, the checker invokes a complete callback.
Error in the complete callback are failure of the user code implementing
the callback, not a checker error, so it is a good idea to handle the
error immediately, instead of propagating the error to the event loop.

When an error in a complete callback is unhandled, we would see this
log:

2017-07-11 23:02:41,420+0300 ERROR (check/loop) [storage.check]
Unhandled error in complete callback (check:339)
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/vdsm/storage/check.py", line 337, in _check_completed
    self._complete(result)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/monitor.py", line 496, in _pathChecked
    raise RuntimeError("Faulty complete callback")
RuntimeError: Faulty complete callback

Change-Id: Ie07f99ab95aa210fb9f33834f980a7e9caeec70e
Bug-Url: https://bugzilla.redhat.com/1469175
Signed-off-by: Nir Soffer <nsoffer@redhat.com>