monitor: _notifyStatusChange should not be a cancelpoint
This method is very fast, so there is no need to make it a cancellation
point. Also, it is called from the checker event loop thread, so it may
abort this thread in the unlikely case when the monitor thread is
stopped while the checker event loop thread is blocked on this method.

Change-Id: Ia962dc845fb790f0a8dc5d296a122d2d33edcf22
Bug-Url: https://bugzilla.redhat.com/1469175
Signed-off-by: Nir Soffer <nsoffer@redhat.com>