RPC: Add missing return statement
In the current code, the function _processEvent()
is not reachable, so this change does not affect anything.
But it will be reachable in subsequent patches.

It is unreachable, because the JsonRpcClient can only
receive messages from a single subscription queue and
this specific queue does not contain events,
only rpc responses

Change-Id: I17a948211ba3aa05af2be69c2a2d366aa26a9e9b
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>