net: NM acquire should ignore duplicate ifcfg files
When quering NM DBUS to translate a file path to a connection ID, an
error with "is not an ifcfg base file" is raised when the file is a copy
of an existing ifcfg file (a backup).

This patch ignores these king of error the same way it did for non ifcfg
files that are named with the ifcfg prefix.

Change-Id: I337db762d7c6f6247894d7f75096328f5726c273
Bug-Url: https://bugzilla.redhat.com/1441530
Signed-off-by: Edward Haas <edwardh@redhat.com>