virt: vm: add try_to_extend_drive
Extract Vm.try_to_extend_drive from Vm.extendDrivesIfNeeded.
The new method will try to extend a single drive, while the
existing method always scanned all the Vm drives, trying to
extend any of them.

We reimplement Vm.extendDrivesIfNeeded on top of the new method,
with no changes in behaviour.

This change makes the room for the usage of the libvirt
BLOCK_THRESHOLD event. When this event fires, we will know
which drive needs to be extended, so it is pointless and wasteful
to scan all the drives again.

Change-Id: I8e8d2f3f554a18c8e6ef5e43ed6261e15a43bd6b
Signed-off-by: Francesco Romani <fromani@redhat.com>