stomp: fix AsyncDispatcher next_check_interval
If next_check_interval returns 0, asyncore loop is triggerred.
In case of no outgoing heartbeat, the function doesn't ignore the
calculation and returns 0, which causes the loop to be triggerred
constantly and send heartbeats.
On server side outgoing heartbeats are always greater than 0 so we
didn't encounter this behavior before. When we added the same
functionality to client side, where outgoing heartbeat is 0 by default,
we encountered high cpu usage.

Change-Id: I4e7fe82f80535ff851eaf909de194419511154ba
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1376843