introduce vdsm.common.constants
Many constants are used by all vdsm subsystems. This patch move
vdsm.constants into vdsm.common in order to reflect this fact.
The patch keeps a proxy module in vdsm.constants, so that we don't have
to touch all the modules that import vdsm.constants in this single
patch.

vdsm.constants contains constants should be private to a specific
module or subpackage. After moved to vdsm.common, I would like to
continue to clean it up, and to keep there only values that must be
shared by multiple different subsystems.

Change-Id: I8568a806f9195007620d15eff59e72b55042557b
Signed-off-by: Dan Kenigsberg <danken@redhat.com>