recovery: Recover persistent domains in down state
Persistent down domains are currently simply undefined.  But that's not
correct, we want to keep them and to report them to Engine until they are
destroyed.  A typical case, besides other less or more corner cases, when a VM
in DOWN state is recovered is when the VM is shutdown or killed while Vdsm is
down.

We need the following changes to recover VMs in down state:

- Don't undefine domains in down state on recovery.
- Permit calling self._dom.state on those domains in order to be able to
  examine their state.
- If exit info is not found in metadata on recovery, set it.

Change-Id: I269426e04817496a5a7c69b702b8c2cb1b1c94d9
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>