tests: virt: add test for Vm.acpi_enabled()
In commit a877434796eeb5af513 we started to use DomainDescriptor
even for VM configured using the legacy vm.conf format.
We want one uniform way to inspect the Vm configuration, regardless
of how the Vm was started.

Unfortunately, the Vm.acpi_enabled() call started to misreport the
ACPI state. This happens because in the Vm.conf flow we build
early a minimal DomainDescriptor out of the configuration we received,
but we don't call all the initialization method we need.

Add tests to demonstrate this issue.

Change-Id: I1f84e3d31702380162e39686322398a83d326917
Bug-Url: https://bugzilla.redhat.com/1472286
Signed-off-by: Francesco Romani <fromani@redhat.com>