virt: libvirtxml: don't do appendMetadata in init
The __init__ method of libvirtxml.Domain claims that it will
create a minimal 'skeleton Domain'.
If this is the purpose, then the method does already too much:
appendMetadata doesn't belong here, so this patch removes from it.

We should do the least possible amount of work in __init__, and
if we add metadata in here it becomes hard to draw a line to
distinguish what belongs and what doesn't.

With this patch, we aim to establish a boundary: anything that is
currently handled by methods (e.g. appendOs, appendMetadata...)
should not be handled in __init__, which will now only contain the
very basic mandatory elements (e.g. memory, vcpus).

On the other hand, there is the need for a convenient way to create
a simple valid Domain to represent the VM configuration.
This means that we will add all the elements that don't require
setup or non-trivial logic, ruling out hugepages, numa settings, devices.
We will use this in the Vm class to uniformly access the Vm
configuration regardless if it is given in the old vm.conf format
or in the new domain XML format.

This is going to be handled by one upcoming patch will add one
helper to make a minimal domain, calling all the right methods.

With all the patches in this topic, we will have:
1. call libvirtxml.Domain(...) only to get the skeleton, the bare
minimum domain representation.
2. call the new helper method to get a minimal but meaningful domain
representation
3. call the various append* methods to complete the representation,
adding extra elements which require extra setup steps.

Change-Id: I8c08692cf2df0fc54a861c21ec8d81dc415408a7
Bug-Url: https://bugzilla.redhat.com/1472286
Signed-off-by: Francesco Romani <fromani@redhat.com>