kvm2ovirt: change storage types to volume and path
Change how we pass source disks to kvm2ovirt. Instead of making the
distinction based on the disk type -- file vs. block device --
differentiate between the methods we use to access the disk, i.e. volume
in storage pool vs. local path.

As previously with 'file' and 'block' methods the new methods 'path' and
'volume' are mutualy exclusive. For 'path' we expect path to the disk
file or block device. For 'volume' we pass libvirt volume key (which is
basicaly also path).

Change-Id: Ia5a8646a28c3d399f392d3df3508511f904f8383
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>