vm: use metadata for storage configuration
In order to make it possible the recovery using only the
XML as source of data, and thus get rid of the recovery files,
we must make sure the storage configuration is included in
the domain XML.

At the moment this is not true, because the storage configuration
- and only that - is still being sent in the legacy format.

Until the storage support matures, the quickest way out is just
move all this data in the domain metadata. We add some overhead,
but we unlock the way forward.

Change-Id: I9b1b1680fe0c4fc7722a44477dace84ac826dad2
Signed-off-by: Francesco Romani <fromani@redhat.com>