fileVolume: Fix volume path spliting when domain or server path contains "images"
In getChildren helper that was introduced in commit 317b957f (New
getChildrenList implementation), we get the domain path from the
image path by splitting the path by 'images'. However, this
logic is broken if the path contains 'images'. For example:

    /rhev/data-center/mnt/images.com:_path/...
    /rhev/data-center/mnt/server:_images/...

In this patch, we improved the split logic by:
1. Split by '/images/' instead of 'images'
2. Split from right instead of from left

In addition, as the same split logic is implemented in
getDomUuidFromVolumePath helper, we fixed it as well.

Change-Id: I7710de12a5d6d2f69bd2f002f92c30006aad90f9
Bug-Url: https://bugzilla.redhat.com/1461029
Signed-off-by: Ala Hino <ahino@redhat.com>