metadata: add support for complex storage metadata
We want to copy the storage configuration from the legacy format
to the domain metadata as quick fix to unlock the engine-XML based
recovery, while the full support matures.

To do so, we must make sure the existing code in metadata.py can
cope with the nested sections required by storage device:

- 'address'
- 'specParams'
- 'vm_custom'
- 'volumeChain'
- 'volumeInfo'

This patch implements this support.

Change-Id: I395342f658968b627d30a4aa7d50d47a0418fc8a
Signed-off-by: Francesco Romani <fromani@redhat.com>