ssl: handle handshake errors
There are two possible errors (sslerror and socket.error) which are
raised during ssl handshake that need to be handled. We do not want
to see unhandled error backtraces in the logs which could easily
generate enourmos amount of logs when malicious client exploit it.
We log now only the occurence of the error and the offending IP address
that caused the error.

We fix the issue by handling both errors and cleanly closing file
descriptor.

Bug-url: https://bugzilla.redhat.com/1473295
Change-Id: I99cfa35e608f429640455c35495be1783854e3da
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Signed-off-by: Irit Goihman <igoihman@redhat.com>