asyncevent: Improve asyncore integration
Add poll() function based on asyncore.poll2() from python 3.7, and 2.7.
This function waits for events on the event loop channels, and return
list of ready dispatchers, instead of invoking the I/O callbacks
immediately, which is racy.

The ready dispatchers are added to the event loop ready queue, and the
callbacks are invoked at the end of _run_once, in the same way it is
done in the python 3 event loop.

With this change, a faulty handle_error() raising exception will not
cause the event loop to skip the next I/O callbacks. The error will be
logged and then the next callbacks will be handled.

This change also fix the following python bugs that may never be fixed
in 2.7 or even in 3.x, since asyncore was deprecated since version 3.6.

- http://bugs.python.org/issue30994
- http://bugs.python.org/issue30931

Change-Id: Ie047f083624f58b6d7b55970a52492f77d69dcf1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>