tests: Document how to run pytests as root
pytest is trying to keep the last 3 runs temporary directories, to allow
inspection of test files after failed tests. This does not work well
when running the tests both as root and regular user, when pytest
running as regular user is trying to clean up old temporary directories
created by root.

The recommended way to solve this issue is to use separte --basetemp
directories for root and regular user. Add section in the README about
running the tests as root.

Change-Id: Id5b7f984a59a374fcfcdb60976e5685891af3f60
Signed-off-by: Nir Soffer <nsoffer@redhat.com>