storage: Handle inaccessible storage when creating a drive
We used to initialize diskType in __init__, to tell if it is a file or
block storage. However this may fail during recovery if the storage is
not available when creating the drive. Make diskType a computed property
so we can create a drive at this point. Any operation that needs the
disk type will fail later with OSError, until storage becomes accessible
again.

Change-Id: I9513c1c033f7b2e860d72af7a9244b533a1ff1fb
Bug-Url: https://bugzilla.redhat.com/1022961
Signed-off-by: Nir Soffer <nsoffer@redhat.com>