virt: metadata: devices: more attribs for lookup
we need a way to match per-device metadata to its device.
The current code does that (ab)using the `alias' attribute.
Engine -or any other client- is free to add `alias' attributes
to the domain XML, so Vdsm could use this to match metadata
with devices.

Libvirt doesn't allow the client to set aliases, so those
aliases will be discarded.
This works fine on VM creation, but will make accessing the metadata
on recovery impossible: the alias will be overwritten by libvirt, so any
metching will fail.

While we look for a better approach, we solve this issue for the
network devices trivially using the mac address, which is non-ambiguous,
already used by Engine, and guaranteed to be present in all the cases
we care about.

Storage device could follow suit using the "path" attribute.
This patch also adds tests demonstrating when the match fail and when
it will succeed.

Change-Id: Ibbbe94b7081bea5458ff3b9650835405d742e7b5
Signed-off-by: Francesco Romani <fromani@redhat.com>