logging: Handle exception logging when calling getVolumeInfo
Until now, when getVolumeInfo is called and the volume didn't exist, a
VolumeDoesNotExist is raised. Raising that exception when asking for the
volume info is misleading and indicating that something is wrong in the
system, while it is legitimate to ask for the volume info for a volume
that doesn't exist. In this patch, we mark the VolumeDoesNotExist error
as expected when calling getVolumeInfo.

Change-Id: I55f7451078063312fa0cf29021def72ca4856750
Bug-Url: https://bugzilla.redhat.com/1408304
Signed-off-by: Ala Hino <ahino@redhat.com>