virt: Store block job info to metadata
libvirt doesn't provide means to get information about running block jobs so we
must store it to metadata.  The stored information is a nested dictionary and
the default Vdsm metadata mechanism doesn't provide means to handle that.  We
must use a custom way to store the info and we (ab)use JSON for that rather
than trying to manage it specifically in metadata.py, which would hardly bring
any real benefit.

Change-Id: Ibf7312ad524e72010698b024d0524c2271db93e7
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>