net: Centralize call to pre-config validation
Introduce network.validator module, which is responsible to call all
pre-config validators.

This patch creats a new validator module for the link package and moves
the call to these validations (bond naming) to the new network.validator
module.

New pre-config validations should be called from this new module and
existing validation calls should be migrated and called from the new
module as well.

Change-Id: Icd8fab6168816cd03545f268f8402f06e1fe84d6
Signed-off-by: Edward Haas <edwardh@redhat.com>