virt: Never leave a recovered VM in WAIT_FOR_LAUNCH status
If we recover a VM and get an unexpected VM state from libvirt (which can
happen only in unexpected situations we are not aware of), the VM remains in
WAIT_FOR_LAUNCH status.  That means clientIF._waitForDomainsUp will wait
forever and Vdsm will remain unresponsive in recovery state.

To fix that problem we set the VM status to UP (as the best effort fallback
consistent with libvirt inaccessibility handling) in such a case rather than
leaving it in WAIT_FOR_LAUNCH.

Change-Id: I94327e92a09816073ce1aece2479bed08fff5326
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>