virt: Remove stale domains with the same name
When we start a new VM we look for its old stale domain by the VM id.  If
found, the stale domain is removed.  However libvirt also refuses to run a VM
with a different id but the same name as an existent domain.  So we should also
remove stale domains with the same name.

A stale domain with the same name and a different id may occur when the host
crashes, leaving the stale domain behind, and the user removes and creates
again the original VM in Engine.

Change-Id: I5b84a43dc5718f26b52a371a324f84cd5bb3c60c
Bug-Url: https://bugzilla.redhat.com/1477151
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>