virt: API: move api.logged from vm to API.py
Using the @api.logged decorator in our API.py module
leads to a cleaner architecture and to slightly cleaner logs,
especially true for the VM.create case, which has misleading
log for the START.

We also change the logger name from "vdsm.api" to "api.{virt,host}",
depending on the specific API being decorated.
The reason for this change is that the original logger name was
too generic and not really well designed in the first place.

Signed-off-by: Francesco Romani <fromani@redhat.com>
Change-Id: Icd01938909feb285cd1649f5cc83d13726103e51