net: Hide zero-mac nics only if they are identified as vmfex
zero-mac interfaces which have been identified as VF have been assumed
to be vmfex interfaces.
There seem to be other VF interfaces which appear with zero-mac that are
not vmfex.

To make it more specific, the vmfex driver name is examined ('enic').
zero-mac interfaces that are not vmfex are now reported as nics by the
network caps.

Change-Id: I14bca017c9018c7ff50d9cb84311d0d5d655ec3d
Bug-Url: https://bugzilla.redhat.com/1474638
Signed-off-by: Edward Haas <edwardh@redhat.com>