virt: metadata: more generic drive attr matching
In order to support various disk types (e.g. LUN devices)
we need to generalize the attribute matching we do.

We need to match a drive with its metadata, but we also
need to use the minimal set of attributes.

We try to use GUID first, and the pair iface/index if this fails.
This covers all the known cases, even though will need
further work, e.g. using the path if available.

Change-Id: I3cd8f3c5d8b421d0bc573dd559446f09f6ba53ec
Signed-off-by: Francesco Romani <fromani@redhat.com>