virt: deprecate some memory statistics
Deprecate some memory statistics returned by ovirt-guest-agent. The
rationale is that we plan to switch gathering of statistics from
ovirt-guest-agent to balloon driver memory statistics and not all values
can be reported that way.

swap_total, swap_usage: these stats can be happily removed right away
without breaking Engine or MOM.

mem_buffers, mem_cached: these stats are marked as deprected but we will
still stend them (if OGA is installed in guest) because they are part of
Engine REST API. The plan is to get the values properly via balloon
driver. If we succeed in doing that we can drop the deprecation. If not,
we'll have to deprecate the values in REST API too.

Finaly, there is the 'mem_free' stat which is problematic. It cannot be
removed now because both Engine and MOM code slightly depends on it and
have to be patched before we can remove the stat completely. For Linux
VMs it used to contain sum:

       mem_unused + mem_buffers + mem_cached

For now the stat will return same value as mem_unused which is just the
free memory without any reclaimable memory included.

Change-Id: I8577d94a3fdab84a39a2b251cde5b0f63f4db12b
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>