net: Wait for vlan link creation when bond-vlan-dhcp is used
In a setup with a bridgeless network that has bond, vlan and DHCP, the
ifup action is performed in a nonblocking manner and may cause a race
where the vlan link is assumed to exist but the ifup has not yet
executed.

This patch introduces a wait-for-link mechanism for this specific scenario.

Change-Id: I3b638ec5b3fe974be04d3e62437553262978f587
Bug-Url: https://bugzilla.redhat.com/1479829
Signed-off-by: Edward Haas <edwardh@redhat.com>