virt: replace page fault counter stat
This change deprecates the 'pageflt' counter, representing total number
of page faults, and adds new counter 'minflts', representing number of
minor page faults. Together with the existing counter for major page
faults, the two are able to replace the 'pageflt' counter.

The counter for total number of page faults does not exist by itself and
VDSM (or the guest agent before) has to compute the value from the sum
of major and minor page faults. Later, MOM performs the subtraction of
major page faults to again recover the count of minor page faults.
Having the separate counters for major and minor page faults removes the
need for this redundant arithmetic exercise.

Change-Id: Ic2cc421b14f23b4c9a69659c892919af955a813d
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>